+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # HEAD
 
+-   Add `LocalInsDb.merge()` method to merge two databases [#7](https://github.com/ziotom78/libinsdb/pull/7)
+
 # Version 0.5.0
 
 -   Support schema.json.gz, schema.yaml, schema.yaml.gz files [#6](https://github.com/ziotom78/libinsdb/pull/6)